+++ src/main/java/com/truelayer/java/mandates/entities/mandate/VRPSweepingMandate.java
@@ -2,8 +2,8 @@ package com.truelayer.java.mandates.entities.mandate;
 
 import static com.truelayer.java.mandates.entities.mandate.Mandate.Type.SWEEPING;
 
-import com.truelayer.java.entities.providerselection.ProviderSelection;
 import com.truelayer.java.mandates.entities.beneficiary.Beneficiary;
+import com.truelayer.java.payments.entities.providerselection.ProviderSelection;
 import lombok.Builder;
 import lombok.Getter;
