+++ src/main/java/com/truelayer/java/mandates/entities/mandatedetail/MandateDetail.java
@@ -5,9 +5,9 @@ import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.truelayer.java.TrueLayerException;
 import com.truelayer.java.entities.CurrencyCode;
 import com.truelayer.java.entities.User;
-import com.truelayer.java.entities.providerselection.ProviderSelection;
 import com.truelayer.java.mandates.entities.Constraints;
 import com.truelayer.java.mandates.entities.beneficiary.Beneficiary;
+import com.truelayer.java.payments.entities.providerselection.ProviderSelection;
 import java.time.ZonedDateTime;
 import java.util.Map;
 import lombok.*;
