+++ src/main/java/com/truelayer/java/payments/entities/paymentmethod/BankTransfer.java
@@ -2,8 +2,8 @@ package com.truelayer.java.payments.entities.paymentmethod;
 
 import static com.truelayer.java.payments.entities.paymentmethod.PaymentMethod.Type.BANK_TRANSFER;
 
-import com.truelayer.java.entities.providerselection.ProviderSelection;
 import com.truelayer.java.payments.entities.beneficiary.Beneficiary;
+import com.truelayer.java.payments.entities.providerselection.ProviderSelection;
 import lombok.Builder;
 import lombok.Getter;
