+++ src/main/java/com/truelayer/java/payments/entities/providerselection/PreselectedProviderSelection.java
@@ -1,7 +1,6 @@
 package com.truelayer.java.payments.entities.providerselection;
 
 import com.truelayer.java.entities.Remitter;
-import com.truelayer.java.entities.providerselection.ProviderSelection;
 import com.truelayer.java.payments.entities.SchemeId;
 import lombok.Builder;
 import lombok.Getter;
