+++ src/test/java/com/truelayer/java/acceptance/MandatesAcceptanceTests.java
@@ -12,7 +12,6 @@ import com.truelayer.java.entities.CurrencyCode;
 import com.truelayer.java.entities.ProviderFilter;
 import com.truelayer.java.entities.User;
 import com.truelayer.java.entities.accountidentifier.AccountIdentifier;
-import com.truelayer.java.entities.providerselection.ProviderSelection;
 import com.truelayer.java.http.entities.ApiResponse;
 import com.truelayer.java.mandates.entities.*;
 import com.truelayer.java.mandates.entities.Constraints.PeriodicLimits;
@@ -23,6 +22,7 @@ import com.truelayer.java.mandates.entities.mandatedetail.MandateDetail;
 import com.truelayer.java.mandates.entities.mandatedetail.Status;
 import com.truelayer.java.payments.entities.*;
 import com.truelayer.java.payments.entities.paymentmethod.PaymentMethod;
+import com.truelayer.java.payments.entities.providerselection.ProviderSelection;
 import java.net.URI;
 import java.util.Collections;
 import java.util.Objects;
