+++ src/test/java/com/truelayer/java/acceptance/PaymentsAcceptanceTests.java
@@ -9,8 +9,6 @@ import com.truelayer.java.commonapi.entities.SubmitPaymentReturnParametersRespon
 import com.truelayer.java.entities.*;
 import com.truelayer.java.entities.Address;
 import com.truelayer.java.entities.accountidentifier.SortCodeAccountNumberAccountIdentifier;
-import com.truelayer.java.entities.providerselection.ProviderSelection;
-import com.truelayer.java.entities.providerselection.UserSelectedProviderSelection;
 import com.truelayer.java.http.entities.ApiResponse;
 import com.truelayer.java.merchantaccounts.entities.MerchantAccount;
 import com.truelayer.java.payments.entities.*;
@@ -20,6 +18,8 @@ import com.truelayer.java.payments.entities.paymentdetail.PaymentDetail;
 import com.truelayer.java.payments.entities.paymentdetail.forminput.Input;
 import com.truelayer.java.payments.entities.paymentmethod.PaymentMethod;
 import com.truelayer.java.payments.entities.providerselection.PreselectedProviderSelection;
+import com.truelayer.java.payments.entities.providerselection.ProviderSelection;
+import com.truelayer.java.payments.entities.providerselection.UserSelectedProviderSelection;
 import java.net.URI;
 import java.time.LocalDate;
 import java.util.*;
